Restaurant (part 20). Start.

+++ 3.JavaMultithreading/src/com/codegym/task/task27/task2712/ManagerTablet.java
@@ -15,7 +15,7 @@ public class ManagerTablet {
         Collections.sort(list);
 
         for (String key : list) {
-            double aLong = profitMap.get(key);
+            long aLong = profitMap.get(key);
             System.out.println(key + " - " + (aLong / 100) + "." + (aLong % 100));
         }
     }
